fix memheap void pointer arithmetics Refs trac:3763
This commit was SVN r29856.

The following Trac tickets were found above:
  Ticket 3763 --> https://svn.open-mpi.org/trac/ompi/ticket/3763

+++ oshmem/mca/memheap/base/memheap_base_static.c
@@ -62,12 +62,12 @@ int mca_memheap_base_static_init(mca_memheap_map_t *map)
             s->shmid = MEMHEAP_SHM_INVALID;
             s->start = memheap_context.mem_segs[i].start;
             s->end = memheap_context.mem_segs[i].end;
-            s->size = s->end - s->start;
+            s->size = ((uintptr_t)s->end - (uintptr_t)s->start);
             s->type = MAP_SEGMENT_STATIC;
             s->context = NULL;
             map->n_segments++;
 
-            total_mem += s->end - s->start;
+            total_mem += ((uintptr_t)s->end - (uintptr_t)s->start);
         }
         MEMHEAP_VERBOSE(1,
                         "Memheap static memory: %llu byte(s), %d segments",
@@ -177,12 +177,12 @@ static int _load_segments(void)
         memset(&seg, 0, sizeof(seg));
         sscanf(line,
                "%llx-%llx %s %llx %s %llx %s",
-               (long long *) &seg.start,
-               (long long *) &seg.end,
+               (unsigned long long *) &seg.start,
+               (unsigned long long *) &seg.end,
                seg.perms,
-               (long long *) &seg.offset,
+               (unsigned long long *) &seg.offset,
                seg.dev,
-               (long long *) &seg.inode,
+               (unsigned long long *) &seg.inode,
                seg.pathname);
 
         if (OSHMEM_ERROR == _check_address(&seg))
